Added basic support for type tests and type guards on numeric types
The expression "expr IS Type" where Type and the type of expr are both numeric types returns whether the value of expr can be represented by Type. It is equivalent to "Type(expr) = expr".

The expression "expr(Type)" converts expr to type if is safe to do so and fails otherwise. It is equivalent to "ASSERT (expr IS Type); RETURN Type(expr);"


git-svn-id: https://svn-dept.inf.ethz.ch/svn/lecturers/a2/trunk@8976 8c9fc860-2736-0410-a75d-ab315db34111